@@ -1,14 +1,14 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Furion" Version="2.9.2" />
-    <PackageReference Include="Furion.Extras.Authentication.JwtBearer" Version="2.9.2" />
-    <PackageReference Include="Furion.Extras.Logging.Serilog" Version="2.9.2" />
-    <PackageReference Include="Furion.Extras.ObjectMapper.Mapster" Version="2.9.2" />
+    <PackageReference Include="Furion" Version="3.1.3" />
+    <PackageReference Include="Furion.Extras.Authentication.JwtBearer" Version="3.1.3" />
+    <PackageReference Include="Furion.Extras.Logging.Serilog" Version="3.1.3" />
+    <PackageReference Include="Furion.Extras.ObjectMapper.Mapster" Version="3.1.3" />
   </ItemGroup>
 
 </Project>
