@@ -8,7 +8,7 @@
     :class="[prefixCls, theme, { 'collapsed-show-title': getCollapsedShowTitle }]"
     @click="handleGoHome"
   >
-    <img src="../../../assets/images/logo.png" />
+    <img src="../../../assets/svg/logo.svg" />
     <div
       class="ml-2 truncate md:opacity-100"
       :class="[
