@@ -64,184 +64,180 @@
   </div>
 </template>
 <script lang="ts">
-  import { defineComponent, computed } from 'vue';
+import { defineComponent, computed } from 'vue';
 
-  import { AppLogo } from '/@/components/Application';
-  import { AppLocalePicker, AppDarkModeToggle } from '/@/components/Application';
-  import LoginForm from './LoginForm.vue';
-  import ForgetPasswordForm from './ForgetPasswordForm.vue';
-  import RegisterForm from './RegisterForm.vue';
-  import MobileForm from './MobileForm.vue';
-  import QrCodeForm from './QrCodeForm.vue';
+import { AppLogo } from '/@/components/Application';
+import { AppLocalePicker, AppDarkModeToggle } from '/@/components/Application';
+import LoginForm from './LoginForm.vue';
+import ForgetPasswordForm from './ForgetPasswordForm.vue';
+import RegisterForm from './RegisterForm.vue';
+import MobileForm from './MobileForm.vue';
+import QrCodeForm from './QrCodeForm.vue';
 
-  import { useGlobSetting } from '/@/hooks/setting';
-  import { useI18n } from '/@/hooks/web/useI18n';
-  import { useDesign } from '/@/hooks/web/useDesign';
-  import { useLocaleStore } from '/@/store/modules/locale';
+import { useGlobSetting } from '/@/hooks/setting';
+import { useI18n } from '/@/hooks/web/useI18n';
+import { useDesign } from '/@/hooks/web/useDesign';
+import { useLocaleStore } from '/@/store/modules/locale';
 
-  export default defineComponent({
-    name: 'Login',
-    components: {
-      AppLogo,
-      LoginForm,
-      ForgetPasswordForm,
-      RegisterForm,
-      MobileForm,
-      QrCodeForm,
-      AppLocalePicker,
-      AppDarkModeToggle,
+export default defineComponent({
+  name: 'Login',
+  components: {
+    AppLogo,
+    LoginForm,
+    ForgetPasswordForm,
+    RegisterForm,
+    MobileForm,
+    QrCodeForm,
+    AppLocalePicker,
+    AppDarkModeToggle,
+  },
+  props: {
+    sessionTimeout: {
+      type: Boolean,
     },
-    props: {
-      sessionTimeout: {
-        type: Boolean,
-      },
-    },
-    setup() {
-      const globSetting = useGlobSetting();
-      const { prefixCls } = useDesign('login');
-      const { t } = useI18n();
-      const localeStore = useLocaleStore();
+  },
+  setup() {
+    const globSetting = useGlobSetting();
+    const { prefixCls } = useDesign('login');
+    const { t } = useI18n();
+    const localeStore = useLocaleStore();
 
-      return {
-        t,
-        prefixCls,
-        title: computed(() => globSetting?.title ?? ''),
-        showLocale: localeStore.getShowPicker,
-      };
-    },
-  });
+    return {
+      t,
+      prefixCls,
+      title: computed(() => globSetting?.title ?? ''),
+      showLocale: localeStore.getShowPicker,
+    };
+  },
+});
 </script>
 <style lang="less">
-  @prefix-cls: ~'@{namespace}-login';
-  @logo-prefix-cls: ~'@{namespace}-app-logo';
-  @countdown-prefix-cls: ~'@{namespace}-countdown-input';
-  @dark-bg: #293146;
+@prefix-cls: ~'@{namespace}-login';
+@logo-prefix-cls: ~'@{namespace}-app-logo';
+@countdown-prefix-cls: ~'@{namespace}-countdown-input';
+@dark-bg: #293146;
 
-  html[data-theme='dark'] {
-    .@{prefix-cls} {
-      background-color: @dark-bg;
+html[data-theme='dark'] {
+  .@{prefix-cls} {
+    background-color: @dark-bg;
 
-      &::before {
-        background-image: url(/@/assets/svg/login-bg-dark.svg);
-      }
+    &::before {
+      background-image: url(/@/assets/svg/login-bg-dark.svg);
+    }
 
-      .ant-input,
-      .ant-input-password {
-        background-color: #232a3b;
-      }
+    .ant-input,
+    .ant-input-password {
+      background-color: #232a3b;
+    }
 
-      .ant-btn:not(.ant-btn-link):not(.ant-btn-primary) {
-        border: 1px solid #4a5569;
-      }
+    .ant-btn:not(.ant-btn-link):not(.ant-btn-primary) {
+      border: 1px solid #4a5569;
+    }
 
-      &-form {
-        background: transparent !important;
-      }
+    &-form {
+      background: transparent !important;
+    }
 
-      .app-iconify {
-        color: #fff;
-      }
+    .app-iconify {
+      color: #fff;
+    }
+  }
+}
+
+.@{prefix-cls} {
+  overflow: hidden;
+  @media (max-width: @screen-xl) {
+    background-color: #293146;
+
+    .@{prefix-cls}-form {
+      background-color: #fff;
     }
   }
 
-  .@{prefix-cls} {
-    overflow: hidden;
+  &::before {
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    margin-left: -48%;
+    content: '';
     @media (max-width: @screen-xl) {
-      background-color: #293146;
+      display: none;
+    }
+  }
 
-      .@{prefix-cls}-form {
-        background-color: #fff;
-      }
+  .@{logo-prefix-cls} {
+    position: absolute;
+    top: 12px;
+    height: 30px;
+
+    &__title {
+      font-size: 16px;
+      color: #fff;
     }
 
-    &::before {
-      position: absolute;
-      top: 0;
-      left: 0;
-      width: 100%;
-      height: 100%;
-      margin-left: -48%;
-      background-image: url(/@/assets/svg/login-bg.svg);
-      background-position: 100%;
-      background-repeat: no-repeat;
-      background-size: auto 100%;
-      content: '';
-      @media (max-width: @screen-xl) {
-        display: none;
-      }
+    img {
+      width: 32px;
     }
+  }
 
+  .container {
     .@{logo-prefix-cls} {
-      position: absolute;
-      top: 12px;
-      height: 30px;
+      display: flex;
+      width: 60%;
+      height: 80px;
 
       &__title {
-        font-size: 16px;
+        font-size: 24px;
         color: #fff;
       }
 
       img {
-        width: 32px;
+        width: 48px;
       }
-    }
-
-    .container {
-      .@{logo-prefix-cls} {
-        display: flex;
-        width: 60%;
-        height: 80px;
-
-        &__title {
-          font-size: 24px;
-          color: #fff;
-        }
-
-        img {
-          width: 48px;
-        }
-      }
-    }
-
-    &-sign-in-way {
-      .anticon {
-        font-size: 22px;
-        color: #888;
-        cursor: pointer;
-
-        &:hover {
-          color: @primary-color;
-        }
-      }
-    }
-
-    input:not([type='checkbox']) {
-      min-width: 360px;
-
-      @media (max-width: @screen-xl) {
-        min-width: 320px;
-      }
-
-      @media (max-width: @screen-lg) {
-        min-width: 260px;
-      }
-
-      @media (max-width: @screen-md) {
-        min-width: 240px;
-      }
-
-      @media (max-width: @screen-sm) {
-        min-width: 160px;
-      }
-    }
-
-    .@{countdown-prefix-cls} input {
-      min-width: unset;
-    }
-
-    .ant-divider-inner-text {
-      font-size: 12px;
-      color: @text-color-secondary;
     }
   }
+
+  &-sign-in-way {
+    .anticon {
+      font-size: 22px;
+      color: #888;
+      cursor: pointer;
+
+      &:hover {
+        color: @primary-color;
+      }
+    }
+  }
+
+  input:not([type='checkbox']) {
+    min-width: 360px;
+
+    @media (max-width: @screen-xl) {
+      min-width: 320px;
+    }
+
+    @media (max-width: @screen-lg) {
+      min-width: 260px;
+    }
+
+    @media (max-width: @screen-md) {
+      min-width: 240px;
+    }
+
+    @media (max-width: @screen-sm) {
+      min-width: 160px;
+    }
+  }
+
+  .@{countdown-prefix-cls} input {
+    min-width: unset;
+  }
+
+  .ant-divider-inner-text {
+    font-size: 12px;
+    color: @text-color-secondary;
+  }
+}
 </style>
